Split guest span helpers for readability
