+++ crates/temper-wasm/src/engine/guest_spans.rs
@@ -3,15 +3,23 @@ use std::collections::BTreeMap;
 use std::time::SystemTime;
 
 use opentelemetry::KeyValue;
-use opentelemetry::trace::{
-    Event, Span as OtelSpan, SpanId, SpanKind, Status, TraceContextExt as _, TraceId, Tracer,
-};
+use opentelemetry::trace::{SpanId, Status, TraceContextExt as _, TraceId};
 use serde::Deserialize;
 use serde_json::Value;
 use tracing_opentelemetry::OpenTelemetrySpanExt as _;
 
 use crate::host_trait::{datadog_visible_span_hint_field, truncate_for_span_attr};
 use crate::types::WasmInvocationContext;
+
+mod export;
+#[cfg(test)]
+#[path = "guest_spans_test.rs"]
+mod tests;
+
+use export::{
+    export_manual_span, manual_parent_context, manual_span_attributes, merge_end_status_attributes,
+    status_from_end_payload, tracing_span_ids,
+};
 
 const DEFAULT_MAX_GUEST_SPANS: usize = 256;
 
@@ -359,86 +367,12 @@ fn clear_thread_local_guest_spans() {
     ENTERED_GUEST_SPANS.with(|guards| guards.borrow_mut().clear());
 }
 
-fn manual_parent_context() -> Option<opentelemetry::Context> {
-    let context = tracing::Span::current().context();
-    let span_context = context.span().span_context().clone();
-    span_context
-        .is_valid()
-        .then(|| opentelemetry::Context::new().with_remote_span_context(span_context))
-}
-
-fn tracing_span_ids(span: &tracing::Span) -> (Option<TraceId>, Option<SpanId>) {
-    let context = span.context();
-    let span_context = context.span().span_context().clone();
-    if span_context.is_valid() {
-        (Some(span_context.trace_id()), Some(span_context.span_id()))
-    } else {
-        (None, None)
-    }
-}
-
 fn allowed_attributes(attributes: &BTreeMap<String, Value>) -> BTreeMap<String, Value> {
     attributes
         .iter()
         .filter(|(key, _)| guest_span_attribute_allowed(key))
         .map(|(key, value)| (key.clone(), value.clone()))
         .collect()
-}
-
-fn manual_span_attributes(
-    context: &WasmInvocationContext,
-    span_id: i64,
-    attributes: &BTreeMap<String, Value>,
-) -> BTreeMap<String, Value> {
-    let mut manual = allowed_attributes(attributes);
-    manual.insert("tenant".to_string(), Value::String(context.tenant.clone()));
-    manual.insert(
-        "entity_type".to_string(),
-        Value::String(context.entity_type.clone()),
-    );
-    manual.insert(
-        "entity_id".to_string(),
-        Value::String(context.entity_id.clone()),
-    );
-    manual.insert(
-        "trigger_action".to_string(),
-        Value::String(context.trigger_action.clone()),
-    );
-    manual.insert(
-        "wasm_module".to_string(),
-        Value::String(context.wasm_module.clone().unwrap_or_default()),
-    );
-    manual.insert(
-        "agent_id".to_string(),
-        Value::String(context.agent_id.clone().unwrap_or_default()),
-    );
-    manual.insert(
-        "session_id".to_string(),
-        Value::String(context.session_id.clone().unwrap_or_default()),
-    );
-    manual.insert(
-        "workflow_root_entity_type".to_string(),
-        Value::String(
-            context
-                .workflow_root_entity_type
-                .clone()
-                .unwrap_or_default(),
-        ),
-    );
-    manual.insert(
-        "workflow_root_entity_id".to_string(),
-        Value::String(context.workflow_root_entity_id.clone().unwrap_or_default()),
-    );
-    manual.insert(
-        "workflow_run_id".to_string(),
-        Value::String(context.workflow_run_id.clone().unwrap_or_default()),
-    );
-    manual.insert(
-        "observability_event".to_string(),
-        Value::String("wasm_guest.span".to_string()),
-    );
-    manual.insert("wasm_guest_span_id".to_string(), Value::from(span_id));
-    manual
 }
 
 fn merge_allowed_attributes(
@@ -449,103 +383,6 @@ fn merge_allowed_attributes(
         if guest_span_attribute_allowed(key) {
             target.insert(key.clone(), value.clone());
         }
-    }
-}
-
-fn merge_end_status_attributes(
-    target: &mut BTreeMap<String, Value>,
-    payload: &GuestSpanEndPayload,
-) {
-    let status = payload.status.as_deref().unwrap_or("ok");
-    if !status.eq_ignore_ascii_case("error") {
-        return;
-    }
-    let error_message = payload
-        .error_message
-        .as_deref()
-        .filter(|value| !value.is_empty())
-        .unwrap_or("guest span failed");
-    let error_type = payload
-        .error_type
-        .as_deref()
-        .filter(|value| !value.is_empty())
-        .unwrap_or("guest_span_error");
-    target.insert(
-        "error.type".to_string(),
-        Value::String(error_type.to_string()),
-    );
-    target.insert(
-        "error.message".to_string(),
-        Value::String(error_message.to_string()),
-    );
-    target.insert(
-        "exception.message".to_string(),
-        Value::String(error_message.to_string()),
-    );
-}
-
-fn status_from_end_payload(payload: &GuestSpanEndPayload) -> Status {
-    let status = payload.status.as_deref().unwrap_or("ok");
-    if status.eq_ignore_ascii_case("error") {
-        let error_message = payload
-            .error_message
-            .as_deref()
-            .filter(|value| !value.is_empty())
-            .unwrap_or("guest span failed");
-        Status::error(error_message.to_string())
-    } else {
-        Status::Ok
-    }
-}
-
-fn export_manual_span(entry: &GuestSpanEntry, status: Status, end_time: SystemTime) {
-    let (Some(trace_id), Some(span_id), Some(parent_context)) =
-        (entry.trace_id, entry.span_id, entry.parent_context.clone())
-    else {
-        return;
-    };
-
-    let tracer = opentelemetry::global::tracer("temper-wasm-guest");
-    let attrs = entry
-        .attributes
-        .iter()
-        .filter_map(|(key, value)| key_value_from_json(key, value))
-        .collect::<Vec<_>>();
-    let events = entry
-        .events
-        .iter()
-        .map(|event| {
-            let attrs = event
-                .attributes
-                .iter()
-                .filter_map(|(key, value)| key_value_from_json(key, value))
-                .collect::<Vec<_>>();
-            Event::new(event.name.clone(), event.timestamp, attrs, 0)
-        })
-        .collect::<Vec<_>>();
-    let mut span = tracer.build_with_context(
-        tracer
-            .span_builder(entry.name.clone())
-            .with_trace_id(trace_id)
-            .with_span_id(span_id)
-            .with_kind(span_kind(&entry.kind))
-            .with_start_time(entry.start_time)
-            .with_end_time(end_time)
-            .with_status(status)
-            .with_attributes(attrs)
-            .with_events(events),
-        &parent_context,
-    );
-    span.end_with_timestamp(end_time);
-}
-
-fn span_kind(kind: &str) -> SpanKind {
-    match kind.to_ascii_lowercase().as_str() {
-        "server" => SpanKind::Server,
-        "client" => SpanKind::Client,
-        "producer" => SpanKind::Producer,
-        "consumer" => SpanKind::Consumer,
-        _ => SpanKind::Internal,
     }
 }
 
@@ -625,130 +462,5 @@ fn span_record_value(value: &Value) -> String {
         Value::Null | Value::Bool(_) | Value::Number(_) | Value::Array(_) | Value::Object(_) => {
             truncate_for_span_attr(&value.to_string())
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn context() -> WasmInvocationContext {
-        WasmInvocationContext {
-            tenant: "test".to_string(),
-            entity_type: "Session".to_string(),
-            entity_id: "s-1".to_string(),
-            trigger_action: "Run".to_string(),
-            wasm_module: Some("test_guest".to_string()),
-            trigger_params: Value::Null,
-            entity_state: Value::Null,
-            agent_id: Some("agent-1".to_string()),
-            session_id: Some("session-1".to_string()),
-            integration_config: BTreeMap::new(),
-            trace_id: "00000000000000000000000000000001".to_string(),
-            workflow_root_entity_type: Some("Session".to_string()),
-            workflow_root_entity_id: Some("s-1".to_string()),
-            workflow_run_id: Some("Session:s-1".to_string()),
-            http_request: None,
-        }
-    }
-
-    #[test]
-    fn rejects_malformed_and_empty_start_payloads() {
-        let mut registry = GuestSpanRegistry::new(context());
-        assert!(registry.start_span("{").is_err());
-        assert!(registry.start_span(r#"{"name":""}"#).is_err());
-    }
-
-    #[test]
-    fn enforces_nested_lifo_end_order() {
-        let mut registry = GuestSpanRegistry::new(context());
-        let root = registry
-            .start_span(r#"{"name":"guest.root"}"#)
-            .expect("root span should start");
-        let child = registry
-            .start_span(r#"{"name":"guest.child"}"#)
-            .expect("child span should start");
-
-        assert!(registry.end_span(root, r#"{"status":"ok"}"#).is_err());
-        registry
-            .end_span(child, r#"{"status":"ok"}"#)
-            .expect("child can end first");
-        registry
-            .end_span(root, r#"{"status":"ok"}"#)
-            .expect("root can end after child");
-    }
-
-    #[test]
-    fn rejects_unknown_span_ids() {
-        let mut registry = GuestSpanRegistry::new(context());
-        assert!(registry.add_span_event(99, r#"{"name":"event"}"#).is_err());
-        assert!(
-            registry
-                .set_span_attributes(99, r#"{"attributes":{}}"#)
-                .is_err()
-        );
-        assert!(registry.end_span(99, r#"{"status":"ok"}"#).is_err());
-    }
-
-    #[test]
-    fn enforces_per_invocation_span_limit() {
-        let mut registry = GuestSpanRegistry::with_max_spans(context(), 1);
-        registry
-            .start_span(r#"{"name":"guest.root"}"#)
-            .expect("first span should start");
-        assert!(registry.start_span(r#"{"name":"guest.child"}"#).is_err());
-    }
-
-    #[test]
-    fn protects_reserved_observability_fields() {
-        assert!(!guest_span_attribute_allowed("trace_id"));
-        assert!(!guest_span_attribute_allowed("dd.span_id"));
-        assert!(!guest_span_attribute_allowed("_otel.parent_trace_id"));
-        assert!(guest_span_attribute_allowed("tool.name"));
-        assert!(guest_span_attribute_allowed("gen_ai.operation.name"));
-    }
-
-    #[test]
-    fn manual_span_snapshot_includes_invocation_correlation_attributes() {
-        let mut attributes = BTreeMap::new();
-        attributes.insert("tool.name".to_string(), Value::String("python".to_string()));
-        let snapshot = manual_span_attributes(&context(), 7, &attributes);
-
-        assert_eq!(
-            snapshot.get("tenant"),
-            Some(&Value::String("test".to_string()))
-        );
-        assert_eq!(
-            snapshot.get("entity_id"),
-            Some(&Value::String("s-1".to_string()))
-        );
-        assert_eq!(
-            snapshot.get("wasm_module"),
-            Some(&Value::String("test_guest".to_string()))
-        );
-        assert_eq!(
-            snapshot.get("workflow_run_id"),
-            Some(&Value::String("Session:s-1".to_string()))
-        );
-        assert_eq!(
-            snapshot.get("observability_event"),
-            Some(&Value::String("wasm_guest.span".to_string()))
-        );
-        assert_eq!(snapshot.get("wasm_guest_span_id"), Some(&Value::from(7)));
-        assert_eq!(
-            snapshot.get("tool.name"),
-            Some(&Value::String("python".to_string()))
-        );
-    }
-
-    #[test]
-    fn cleanup_closes_unended_spans() {
-        let mut registry = GuestSpanRegistry::new(context());
-        registry
-            .start_span(r#"{"name":"guest.root"}"#)
-            .expect("span should start");
-        assert!(registry.enter_active().is_some());
-        registry.cleanup_unclosed();
-        assert!(registry.enter_active().is_none());
     }
 }
